+++ src/core/include/cesium/omniverse/Context.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "cesium/omniverse/RenderStatistics.h"
 #include "cesium/omniverse/SetDefaultTokenResult.h"
 #include "cesium/omniverse/TokenTroubleshooter.h"
 #include "cesium/omniverse/UsdNotificationHandler.h"
@@ -117,12 +118,23 @@ class Context {
 
     bool creditsAvailable() const;
     std::vector<std::pair<std::string, bool>> getCredits() const;
+    void creditsStartNextFrame();
+
+    RenderStatistics getRenderStatistics() const;
 
   private:
     void processPropertyChanged(const ChangedPrim& changedPrim);
+    void processCesiumDataChanged(const ChangedPrim& changedPrim);
+    void processCesiumTilesetChanged(const ChangedPrim& changedPrim);
+    void processCesiumImageryChanged(const ChangedPrim& changedPrim);
     void processPrimRemoved(const ChangedPrim& changedPrim);
     void processPrimAdded(const ChangedPrim& changedPrim);
     void processUsdNotifications();
+
+    bool getDebugDisableGeometryPool() const;
+    bool getDebugDisableMaterialPool() const;
+    uint64_t getDebugGeometryPoolInitialCapacity() const;
+    uint64_t getDebugMaterialPoolInitialCapacity() const;
 
     std::shared_ptr<TaskProcessor> _taskProcessor;
     std::shared_ptr<HttpAssetAccessor> _httpAssetAccessor;
@@ -148,8 +160,6 @@ class Context {
 
     std::filesystem::path _cesiumExtensionLocation;
     std::filesystem::path _certificatePath;
-
-    bool _debugDisableMaterials{false};
 };
 
 } // namespace cesium::omniverse