+++ src/core/include/cesium/omniverse/FabricMeshManager.h
@@ -45,11 +45,10 @@ class FabricMeshManager {
 
     void releaseMesh(std::shared_ptr<FabricMesh> mesh);
 
-    uint64_t getNumberOfGeometriesInUse() const;
-    uint64_t getNumberOfMaterialsInUse() const;
-
-    uint64_t getGeometryPoolCapacity() const;
-    uint64_t getMaterialPoolCapacity() const;
+    void setDisableGeometryPool(bool disableGeometryPool);
+    void setDisableMaterialPool(bool disableMaterialPool);
+    void setGeometryPoolInitialCapacity(uint64_t geometryPoolInitialCapacity);
+    void setMaterialPoolInitialCapacity(uint64_t materialPoolInitialCapacity);
 
     void clear();
 
@@ -83,8 +82,11 @@ class FabricMeshManager {
     std::vector<std::shared_ptr<FabricGeometryPool>> _geometryPools;
     std::vector<std::shared_ptr<FabricMaterialPool>> _materialPools;
 
-    bool _debugDisableGeometryPool{true};
-    bool _debugDisableMaterialPool{false};
+    bool _disableGeometryPool{false};
+    bool _disableMaterialPool{false};
+
+    uint64_t _geometryPoolInitialCapacity{0};
+    uint64_t _materialPoolInitialCapacity{0};
 
     std::atomic<int64_t> _geometryId{0};
     std::atomic<int64_t> _materialId{0};