+++ src/core/src/FabricUtil.cpp
@@ -500,6 +500,10 @@ std::string printFabricStage() {
 FabricStatistics getStatistics() {
     FabricStatistics statistics;
 
+    if (!UsdUtil::hasStage()) {
+        return statistics;
+    }
+
     auto sip = UsdUtil::getFabricStageInProgress();
 
     const auto geometryBuckets = sip.findPrims(
@@ -516,8 +520,18 @@ FabricStatistics getStatistics() {
 
         auto worldVisibilityFabric =
             sip.getAttributeArrayRd<bool>(geometryBuckets, bucketId, FabricTokens::_worldVisibility);
-        statistics.numberOfGeometriesVisible +=
-            std::count(worldVisibilityFabric.begin(), worldVisibilityFabric.end(), true);
+        auto faceVertexCountsFabric =
+            sip.getArrayAttributeArrayRd<int>(geometryBuckets, bucketId, FabricTokens::faceVertexCounts);
+
+        for (size_t i = 0; i < paths.size(); i++) {
+            const auto triangleCount = faceVertexCountsFabric[i].size();
+            statistics.numberOfTrianglesLoaded += triangleCount;
+
+            if (worldVisibilityFabric[i]) {
+                statistics.numberOfGeometriesVisible++;
+                statistics.numberOfTrianglesVisible += triangleCount;
+            }
+        }
     }
 
     for (size_t bucketId = 0; bucketId < materialBuckets.bucketCount(); bucketId++) {