+++ src/core/src/UsdUtil.cpp
@@ -267,6 +267,11 @@ pxr::CesiumData defineCesiumData(const pxr::SdfPath& path) {
     cesiumData.CreateGeoreferenceOriginLongitudeAttr(pxr::VtValue(-105.25737));
     cesiumData.CreateGeoreferenceOriginLatitudeAttr(pxr::VtValue(39.736401));
     cesiumData.CreateGeoreferenceOriginHeightAttr(pxr::VtValue(2250.0));
+    cesiumData.CreateDebugDisableMaterialsAttr(pxr::VtValue(false));
+    cesiumData.CreateDebugDisableGeometryPoolAttr(pxr::VtValue(true));
+    cesiumData.CreateDebugDisableMaterialPoolAttr(pxr::VtValue(false));
+    cesiumData.CreateDebugGeometryPoolInitialCapacityAttr(pxr::VtValue(uint64_t(0)));
+    cesiumData.CreateDebugMaterialPoolInitialCapacityAttr(pxr::VtValue(uint64_t(2048)));
 
     return cesiumData;
 }
@@ -295,6 +300,7 @@ pxr::CesiumTilesetAPI defineCesiumTileset(const pxr::SdfPath& path) {
     tileset.CreateCulledScreenSpaceErrorAttr(pxr::VtValue(64.0f));
     tileset.CreateSuspendUpdateAttr(pxr::VtValue(false));
     tileset.CreateSmoothNormalsAttr(pxr::VtValue(false));
+    tileset.CreateShowCreditsOnScreenAttr(pxr::VtValue(false));
 
     return tileset;
 }
@@ -306,6 +312,7 @@ pxr::CesiumImagery defineCesiumImagery(const pxr::SdfPath& path) {
 
     imagery.CreateIonAssetIdAttr();
     imagery.CreateIonAccessTokenAttr();
+    imagery.CreateShowCreditsOnScreenAttr(pxr::VtValue(false));
 
     return imagery;
 }